testing/requirements.txt
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-xdist==3.3.1
pytest-json-report==1.5.0
pytest-html==4.1.1
pytest-cov==4.1.0
httpx==0.25.2
websockets==12.0
aiohttp==3.9.1
asyncpg==0.29.0
redis==5.0.1
minio==7.2.0
opencv-python==4.8.1.78
numpy==1.24.4
Pillow==10.1.0
structlog==23.2.0
psutil==5.9.6
PyJWT==2.8.0
cryptography==41.0.8
faker==20.1.0
factory-boy==3.3.0
locust==2.17.0